Append hashes to file names.
File names will now be appended with the first 10 characters of the md5
hash of the file. This ensures that the file names will not collide
while saving them to the disk, because collisions will result in older
files being overwritten. Related tests, and an option to turn this off
if needed has been added as well.