More comprehensive matcher for direct downloader.
Direct downloader used a very simple matcher that essentially just
looked at the end of the URL.

The new matcher also checks if the URL is well-formed. It also accepts
URLs where there is fragment information at the end, such as
`http://example.com/test.jpg#a`, which points to the same resource.