More generic test for hashed imgur albums.
Imgur sends archives with different hashes when downloading an
album. This makes the previous test unreliable, as the hash of the
archive that is appended to the names keeps getting changed.

The current test instead checks for the number of files and a small
sanity check for the name of the files, to make the test more flexible.